pyproject.toml
[project]
name = "live-stranslate"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "livekit-agents[cartesia,gladia,openai]~=1.2",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.3",
]
